+++ packages/pluggableWidgets/datagrid-web/src/Datagrid.editorConfig.ts
@@ -89,7 +89,8 @@ export function getProperties(values: DatagridPreviewProps, defaultProperties: P
             "dynamicPage",
             "dynamicPageSize",
             "useCustomPagination",
-            "customPagination"
+            "customPagination",
+            "totalCountValue"
         ]);
     }