+++ packages/pluggableWidgets/rich-text-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue where hyperlink tag does not read target attribute when edited from code viewer.
+
 ## [4.11.1] - 2026-01-27
 
 ### Fixed